feat: error handler and all controllers

+++ src/main/java/org/example/ativooperante/db/repository/DenunciaRepository.java
@@ -3,4 +3,5 @@ package org.example.ativooperante.db.repository;
 import org.example.ativooperante.db.entities.Denuncia;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DenunciaRepository extends JpaRepository<Denuncia, Integer> { }
+public interface DenunciaRepository extends JpaRepository<Denuncia, Long> {
+}
