+++ src/main/java/org/example/ativooperante/db/repository/FeedbackRepository.java
@@ -3,4 +3,4 @@ package org.example.ativooperante.db.repository;
 import org.example.ativooperante.db.entities.Feedback;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FeedbackRepository extends JpaRepository<Feedback, Integer> { }
+public interface FeedbackRepository extends JpaRepository<Feedback, Long> { }
