+++ src/main/java/org/example/ativooperante/db/repository/OrgaoRepository.java
@@ -2,5 +2,5 @@ package org.example.ativooperante.db.repository;
 import org.example.ativooperante.db.entities.Orgao;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrgaoRepository extends JpaRepository<Orgao, Integer> {
+public interface OrgaoRepository extends JpaRepository<Orgao, Long> {
 }
